+++ src/Processors/Formats/Impl/Parquet/PrepareForWrite.cpp
@@ -169,7 +169,7 @@ void updateRepDefLevelsForArray(ColumnChunkWriteState & s, const IColumn::Offset
         size_t i = 0;
         for (ssize_t row = 0; row < static_cast<ssize_t>(offsets.size()); ++row)
         {
-            size_t n = offsets[row] - offsets[static_cast<ssize_t>(row) - 1];
+            size_t n = offsets[row] - offsets[row - 1];
             if (n)
             {
                 s.rep[i] = 0;
@@ -203,7 +203,7 @@ void updateRepDefLevelsForArray(ColumnChunkWriteState & s, const IColumn::Offset
     size_t empty_arrays = 0;
     for (ssize_t row = 0; row < static_cast<ssize_t>(offsets.size()); ++row)
     {
-        size_t n = offsets[row] - offsets[static_cast<ssize_t>(row) - 1];
+        size_t n = offsets[row] - offsets[row - 1];
         if (n)
         {
             /// Un-increment the first rep of the array.
